+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [3.2.23] - 2022-02-18
 ### Added
 - SectionTitlePrimary and LinkTileButtons moved to Rokwire plugin [#1325](https://github.com/rokwire/illinois-app/issues/1325).
 - FlexContentWidget moved to Rokwire plugin [#1325](https://github.com/rokwire/illinois-app/issues/1325).